+++ presentation/src/main/java/com/sms/presentation/main/ui/fill_out_information/screen/ForeignLanguageScreen.kt
@@ -10,14 +10,12 @@ import com.msg.sms.design.component.spacer.SmsSpacer
 import com.msg.sms.design.component.topbar.TopBarComponent
 import com.msg.sms.design.icon.BackButtonIcon
 import com.sms.presentation.main.ui.fill_out_information.component.ForeignLanguageComponent
-import com.sms.presentation.main.viewmodel.FileUploadViewModel
 import com.sms.presentation.main.viewmodel.FillOutViewModel
 
 @Composable
 fun ForeignLanguageScreen(
     navController: NavController,
-    fillOutViewModel: FillOutViewModel,
-    fileUploadViewModel: FileUploadViewModel
+    viewModel: FillOutViewModel
 ) {
     Column(modifier = Modifier.background(Color.White)) {
         TopBarComponent(text = "정보입력", leftIcon = { BackButtonIcon() }, rightIcon = null) {
@@ -26,8 +24,7 @@ fun ForeignLanguageScreen(
         SmsSpacer()
         ForeignLanguageComponent(
             navController = navController,
-            fillOutViewModel = fillOutViewModel,
-            fileUploadViewModel = fileUploadViewModel
+            viewModel = viewModel
         )
     }
 }
