+++ presentation/src/main/java/com/sms/presentation/main/viewmodel/FillOutViewModel.kt
@@ -5,9 +5,12 @@ import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.msg.sms.domain.model.fileupload.response.FileUploadResponseModel
 import com.msg.sms.domain.model.major.MajorListModel
 import com.msg.sms.domain.model.student.request.CertificateInformationModel
 import com.msg.sms.domain.model.student.request.EnterStudentInformationModel
+import com.msg.sms.domain.usecase.fileupload.DreamBookUploadUseCase
+import com.msg.sms.domain.usecase.fileupload.ImageUploadUseCase
 import com.msg.sms.domain.usecase.major.GetMajorListUseCase
 import com.msg.sms.domain.usecase.student.EnterStudentInformationUseCase
 import com.sms.presentation.main.ui.fill_out_information.data.CertificationData
@@ -22,18 +25,27 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.launch
+import okhttp3.MultipartBody
 import javax.inject.Inject
 
 @HiltViewModel
 class FillOutViewModel @Inject constructor(
     private val enterStudentInformationUseCase: EnterStudentInformationUseCase,
     private val getMajorListUseCase: GetMajorListUseCase,
+    private val imageUploadUseCase: ImageUploadUseCase,
+    private val dreamBookUploadUseCase: DreamBookUploadUseCase
 ) : ViewModel() {
     private val _enterInformationResponse = MutableStateFlow<Event<Unit>>(Event.Loading)
     val enterInformationResponse: StateFlow<Event<Unit>> get() = _enterInformationResponse
 
     private val _getMajorListEvent = MutableStateFlow<Event<MajorListModel>>(Event.Loading)
     val getMajorListEvent = _getMajorListEvent.asStateFlow()
+
+    private val _imageUploadResponse = MutableStateFlow<Event<FileUploadResponseModel>>(Event.Loading)
+    val imageUploadResponse: StateFlow<Event<FileUploadResponseModel>> get() = _imageUploadResponse
+
+    private val _dreamBookUploadResponse = MutableStateFlow<Event<FileUploadResponseModel>>(Event.Loading)
+    val dreamBookUploadResponse: StateFlow<Event<FileUploadResponseModel>> get() = _dreamBookUploadResponse
 
     private val major = mutableStateOf("")
     private val techStack = mutableStateOf("")
@@ -170,6 +182,34 @@ class FillOutViewModel @Inject constructor(
             }
         }.onFailure { error ->
             _enterInformationResponse.value = error.errorHandling()
+        }
+    }
+
+    fun imageUpload(file: MultipartBody.Part) = viewModelScope.launch {
+        imageUploadUseCase(
+            file = file
+        ).onSuccess {
+            it.catch { remoteError ->
+                _imageUploadResponse.value = remoteError.errorHandling()
+            }.collect { response ->
+                _imageUploadResponse.value = Event.Success(data = response)
+            }
+        }.onFailure { error ->
+            _imageUploadResponse.value = error.errorHandling()
+        }
+    }
+
+    fun dreamBookUpload(file: MultipartBody.Part) = viewModelScope.launch {
+        dreamBookUploadUseCase(
+            file = file
+        ).onSuccess {
+            it.catch { remoteError ->
+                _dreamBookUploadResponse.value = remoteError.errorHandling()
+            }.collect { response ->
+                _dreamBookUploadResponse.value = Event.Success(data = response)
+            }
+        }.onFailure { error ->
+            _dreamBookUploadResponse.value = error.errorHandling()
         }
     }
 }
